abstracted generic away into a common directory to be reused by all competition specific sub packages
