+++ toxicity/test/pathmagic.py
@@ -3,6 +3,6 @@
 import os
 import sys
 
-bp = os.path.dirname(os.path.realpath('.')).split(os.sep)
+bp = os.path.dirname(os.path.realpath('')).split(os.sep)
 modpath = os.sep.join(bp)
 sys.path.insert(0, modpath)
