+++ toxicity/utils.py
@@ -1,27 +1,9 @@
-import time
 import pandas as pd
 import numpy as np
 from scipy.sparse import csr_matrix
 from sklearn import preprocessing
 
 TAGS = ['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']
-
-
-def timing(f):
-    """
-    Decorator to time a function call and print results.
-    :param f: Callable to be timed
-    :return: Void. Prints to std:out as a side effect
-    """
-
-    def wrap(*args, **kwargs):
-        start = time.time()
-        ret = f(*args, **kwargs)
-        stop = time.time()
-        print('{} function took {:.1f} seconds to complete\n'.format(f.__name__, (stop - start)))
-        return ret
-
-    return wrap
 
 
 def scale_data(train, test):
